Move Panel outside of api package
