+++ core/shared/src/main/scala/eu/joaocosta/interim/api/Panels.scala
@@ -22,18 +22,6 @@ object Panels extends Panels
 
 trait Panels:
 
-  trait Panel[I, F[_]]:
-    def render[T](area: Ref[PanelState[Rect]], body: I => T): Component[F[T]]
-
-    def apply[T](area: Ref[PanelState[Rect]])(body: I => T): Component[F[T]] =
-      render(area, body)
-
-    def apply[T](area: PanelState[Rect])(body: I => T): Component[F[T]] =
-      render(Ref(area), body)
-
-    def apply[T](area: Rect)(body: I => T): Component[F[T]] =
-      render(Ref(PanelState.open(area)), body)
-
   /**  Window with a title.
     *
     * @param title of this window
